2.- Modificaciones en el return (devuelve un byte en vez de un int).

+++ ExamenEntornos/src/examenentornos.java
@@ -6,7 +6,7 @@ public class examenentornos {
         int b = 7;
         System.out.printf("La suma es de %d + %d es %d",a, b, Sumar(a, b));
     }
-    public static int Sumar(int a, int b){     
-        return a + b;
+    public static byte Sumar(int a, int b){     
+        return (byte)(a + b);
     }
 }
